Move cloud under goaviatrix and fix imports
