+++ aviatrix/data_source_aviatrix_gateway.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/terraform-providers/terraform-provider-aviatrix/cloud"
+	"github.com/terraform-providers/terraform-provider-aviatrix/goaviatrix/cloud"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/terraform-providers/terraform-provider-aviatrix/goaviatrix"
