+++ aviatrix/data_source_aviatrix_vpc_tracker.go
@@ -3,7 +3,7 @@ package aviatrix
 import (
 	"fmt"
 
-	"github.com/terraform-providers/terraform-provider-aviatrix/cloud"
+	"github.com/terraform-providers/terraform-provider-aviatrix/goaviatrix/cloud"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
