+++ aviatrix/resource_aviatrix_transit_gateway.go
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	"github.com/terraform-providers/terraform-provider-aviatrix/cloud"
 	"github.com/terraform-providers/terraform-provider-aviatrix/goaviatrix"
+	"github.com/terraform-providers/terraform-provider-aviatrix/goaviatrix/cloud"
 )
 
 func resourceAviatrixTransitGateway() *schema.Resource {
